Move EventQueue it its own module in h.api
