+++ h/api/__init__.py
@@ -3,6 +3,7 @@
 
 def includeme(config):
     config.include('h.api.db')
+    config.include('h.api.eventqueue')
     config.include('h.api.presenters')
     config.include('h.api.search')
     config.include('h.api.views')
