+++ h/app.py
@@ -47,32 +47,6 @@ def create_app(global_config, **settings):
     return config.make_wsgi_app()
 
 
-class EventQueue(object):
-    def __init__(self, request):
-        self.request = request
-        self.queue = collections.deque()
-
-        request.add_response_callback(self.response_callback)
-
-    def __call__(self, event):
-        self.queue.append(event)
-
-    def publish_all(self):
-        while True:
-            try:
-                event = self.queue.popleft()
-            except IndexError:
-                break
-            self.request.registry.notify(event)
-
-    def response_callback(self, request, response):
-        if request.exception is not None:
-            return
-
-        with request.tm:
-            self.publish_all()
-
-
 def includeme(config):
     config.set_root_factory('h.resources:Root')
 
@@ -119,9 +93,6 @@ def includeme(config):
             "style-src": ["'self'", "fonts.googleapis.com"],
         },
     })
-
-    # After commit events
-    config.add_request_method(EventQueue, name='notify_after_commit', reify=True)
 
     # API module
     #
